Updated files

+++ Back_End/src/main/java/lk/ijse/spring/dto/CarDTO.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.math.BigDecimal;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,12 +25,12 @@ public class CarDTO {
     private String backViewImg;
     private String sideViewImg;
     private String internalViewImg;
-    private double dailyRate;
-    private double monthlyRate;
-    private double freeKmForPrice;
-    private double freeKmForDuration;
-    private double lossDamageWaiver;
-    private double priceForExtraKm;
-    private double completeKm;
+    private BigDecimal dailyRate;
+    private BigDecimal monthlyRate;
+    private BigDecimal freeKmForPrice;
+    private BigDecimal freeKmForDuration;
+    private BigDecimal lossDamageWaiver;
+    private BigDecimal priceForExtraKm;
+    private BigDecimal completeKm;
     private Boolean isAvailable = false;
 }
